// src/app/api/models/BookingCreateDTO.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
import type { EnumTypeHelp } from './EnumTypeHelp';
export type BookingCreateDTO = {
    id?: string;
    slotId?: string | null;
    subject: string;
    description?: string | null;
    typeHelp?: EnumTypeHelp;
};